refs: Fix and refactor argument handling
Unlike for the log and diff prefixes there is no refresh variant
of `magit-show-refs'.  In fact because this command only acts as a
prefix if it is invoked with a prefix argument (a feature that might
be unknown to the user) or when invoked from a refs buffer, it can
be considered to be the refresh variant.  Therefore it should always
use the arguments from the current buffer if it is a refs buffer.