magit-get-mode-buffer: New function
This is the new `magit-mode-get-buffer', but the existing function has
an incompatible signature so we do this in two steps to give authors
time to adapt their code.  Eventually the old version will be removed,
the new renamed to reuse its name, and the temporary name becomes an
obsolete alias.